@@ -1,76 +1,53 @@
 {
-  "name": "ecommerce-seed",
-  "version": "1.0.0",
-  "description": "The E-commerce seed project with Angular2, AngularFire2, and Firebase",
-  "author": {
-    "name": "Sergey Akopkokhyants",
-    "email": "akserg@gmail.com"
-  },
-  "homepage": "https://github.com/akserg/learning_bootstrap_angular2",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/akserg/learning_bootstrap_angular2.git"
-  },
+  "name": "ecommerce",
+  "version": "0.0.0",
   "license": "MIT",
-  "engines": {
-    "node": ">=6.5"
-  },
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack --display-chunks --progress",
-    "clean:coverage": "del-cli ./coverage",
-    "clean:dist": "del-cli ./dist",
-    "lint": "run-s lint:ts",
-    "lint:ts": "tslint './src/**/*.ts'",
-    "prebuild": "run-s clean:dist lint",
-    "server:dev": "cross-env NODE_ENV=development webpack-dev-server",
-    "start": "npm run server:dev"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
+  "private": true,
   "dependencies": {
-    "@angular/common": "2.0.0",
-    "@angular/compiler": "2.0.0",
-    "@angular/core": "2.0.0",
-    "@angular/forms": "2.0.0",
-    "@angular/platform-browser": "2.0.0",
-    "@angular/platform-browser-dynamic": "2.0.0",
-    "@angular/router": "3.0.0",
-    "angularfire2": "2.0.0-beta.5",
-    "core-js": "~2.4.1",
-    "firebase": "3.3.0",
-    "ng2-bootstrap": "^1.1.5",
-    "rxjs": "5.0.0-beta.12",
-    "zone.js": "~0.6.25"
+    "@angular/animations": "^4.0.0",
+    "@angular/common": "^4.0.0",
+    "@angular/compiler": "^4.0.0",
+    "@angular/core": "^4.0.0",
+    "@angular/forms": "^4.0.0",
+    "@angular/http": "^4.0.0",
+    "@angular/platform-browser": "^4.0.0",
+    "@angular/platform-browser-dynamic": "^4.0.0",
+    "@angular/router": "^4.0.0",
+    "angularfire2": "^4.0.0-rc.2",
+    "bootstrap": "^4.0.0-alpha.6",
+    "core-js": "^2.4.1",
+    "firebase": "^4.3.0",
+    "ngx-bootstrap": "^1.9.1",
+    "rxjs": "^5.1.0",
+    "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@types/core-js": "~0.9.33",
-    "@types/node": "~6.0.39",
-    "@types/request": "0.0.30",
-    "autoprefixer": "~6.4.1",
-    "cross-env": "~2.0.1",
-    "css-loader": "^0.25.0",
-    "del-cli": "~0.2.0",
-    "eslint": "~3.5.0",
-    "extract-text-webpack-plugin": "2.0.0-beta.4",
-    "file-loader": "^0.9.0",
-    "firebase-tools": "3.0.7",
-    "html-webpack-plugin": "~2.22.0",
-    "minx": "r-park/minx.git",
-    "morgan": "~1.7.0",
-    "node-sass": "^3.10.0",
-    "nodemon": "~1.10.2",
-    "npm-run-all": "~3.1.0",
-    "postcss-loader": "~0.13.0",
-    "raw-loader": "~0.5.1",
-    "resolve-url-loader": "^1.6.0",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "ts-helpers": "~1.1.1",
-    "ts-loader": "~0.8.2",
-    "tslint": "~3.15.1",
-    "typescript": "~2.0.2",
-    "typings": "~1.3.3",
-    "url-loader": "^0.5.7",
-    "webpack": "2.1.0-beta.22",
-    "webpack-dev-server": "2.1.0-beta.4",
-    "webpack-md5-hash": "~0.0.5"
+    "@angular/cli": "1.1.3",
+    "@angular/compiler-cli": "^4.0.0",
+    "@angular/language-service": "^4.0.0",
+    "@types/jasmine": "2.5.45",
+    "@types/node": "~6.0.60",
+    "@types/request": "^2.0.3",
+    "codelyzer": "~3.0.1",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.1.2",
+    "ts-node": "~3.0.4",
+    "tslint": "~5.3.2",
+    "typescript": "~2.3.3"
   }
 }
