@@ -1,8 +1,9 @@
-import {Injectable} from "@angular/core";
-import {AngularFire, FirebaseListObservable} from "angularfire2";
+import {Injectable} from '@angular/core';
+import {AngularFireDatabase, FirebaseListObservable} from 'angularfire2/database';
 
-import {Observable} from "rxjs/Observable";
-import "rxjs/add/operator/catch";
+import {Observable} from 'rxjs/Observable';
+import 'rxjs/add/operator/catch';
+import 'rxjs/add/observable/throw';
 
 export class Category {
     //  Unique Id
@@ -21,30 +22,34 @@ export class Category {
 export class CategoryService {
 
     // URL to Categories web api
-    private categoriesUrl = "categories";
+    private categoriesUrl = 'categories';
     // We keep categories in cache variable
-    private categories: Category[] = [];
+    private categories: Map<string, Category> = new Map<string, Category>();
 
-    constructor(private af: AngularFire) {}
+    constructor(private db: AngularFireDatabase) { }
 
     getCategories(): Observable<Category[]> {
-        return this.af.database
+        let result = this.db
             .list(this.categoriesUrl)
             .catch(this.handleError);
+
+        result.subscribe((categories: Category[]) => {
+            this.categories = new Map<string, Category>();
+            categories.forEach((category: Category) => {
+                this.categories.set(category.id, category);
+            });
+        });
+        return result
     }
 
     getCategory(id: string): Category {
-        for (let i = 0; i < this.categories.length; i++) {
-            if (this.categories[i].id === id) {
-                return this.categories[i];
-            }
-        }
-        return null;
+        return this.categories.get(id);
     }
 
     private handleError(error: any): Observable<any> {
         let errMsg = (error.message) ? error.message : error.status ?
-            `${error.status} - ${error.statusText}` : "Server error";
+            `${error.status} - ${error.statusText}` : 'Server error';
+        window.alert(`An error occurred: ${errMsg}`);
         return Observable.throw(errMsg);
     }
 }
