@@ -12,8 +12,8 @@ import { AuthService } from "./auth.service";
 
 @Component({
     selector: "db-signin",
-    template: require("./sign-in.component.html"),
-    styles: [require("./auth.scss")]
+    templateUrl: "./sign-in.component.html",
+    styleUrls: ["./auth.scss"]
 })
 export class SignInComponent {
     public form: FormGroup;
@@ -37,12 +37,14 @@ export class SignInComponent {
         this.auth.signIn(values.email, values.password)
             .then(() => this.postSignIn())
             .catch((error) => {
-                this.error = "Username or password is incorrect";
+                this.error = error.toString();
                 this.submitted = false;
             });
     }
 
     private postSignIn(): void {
+        this.email.setValue('');
+        this.password.setValue('');
         this.router.navigate(["/welcome"]);
     }
 }
